third_party/WebKit/LayoutTests/bluetooth/requestDevice/blocklisted-service-in-optionalServices.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  let expected = new DOMException('Origin is not allowed to access the ' +
                                  'service. Tip: Add the service UUID to ' +
                                  '\'optionalServices\' in requestDevice() ' +
                                  'options. https://goo.gl/HxfxSQ',
                                  'SecurityError');
  return getDiscoveredHealthThermometerDevice({
      filters: [{services: ['health_thermometer']}],
      optionalServices: ['human_interface_device']
    })
    .then(({device, fake_peripheral}) => {
      return fake_peripheral
        .setNextGATTConnectionResponse({code: HCI_SUCCESS})
        .then(() => device.gatt.connect())
        .then(() => fake_peripheral.setNextGATTDiscoveryResponse({
          code: HCI_SUCCESS}))
        .then(() => Promise.all([
          assert_promise_rejects_with_message(
            device.gatt.getPrimaryService('human_interface_device'),
            expected, 'Blocklisted service not accessible.'),
          assert_promise_rejects_with_message(
            device.gatt.getPrimaryServices('human_interface_device'),
            expected, 'Blocklisted services not accessible.')]));
    });
}, 'Blocklisted UUID in optionalServices is removed and access not granted.');
</script>
